Uodate icons for dialogs, reorganize update manager and add more tests
